added unit testing for verify age
added unit testing for verify age function. I used node js mocha and chia for the testing and that comes with 300ish node files that I don't think need to be added to the file, if you need to run the unit tests on your machine let me know and I can help you with that.